Fix #584 via repair and upgrade example chapter16 08

+++ examples/chapter16_08/src/mcal_lcd/mcal_lcd_base.h
@@ -1,3 +1,10 @@
+///////////////////////////////////////////////////////////////////////////////
+//  Copyright Christopher Kormanyos 2020 - 2024.
+//  Distributed under the Boost Software License,
+//  Version 1.0. (See accompanying file LICENSE_1_0.txt
+//  or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+
 #ifndef MCAL_LCD_BASE_2020_06_10_H_
   #define MCAL_LCD_BASE_2020_06_10_H_
 
@@ -13,7 +20,7 @@
     virtual bool init() = 0;
 
     virtual bool write(const char* pstr,
-                       const std::uint_fast8_t length,
+                       const std::uint_fast16_t length,
                        const std::uint_fast8_t line_index) = 0;
 
   protected:
